requirements.txt
python-dotenv==0.17.0
keyboard== 0.13.5